requirements.txt
Flask
openai
python-telegram-bot
sqlite3
